Moved three methods ‘serializeKnowledge’,  ‘serializeKB’, ‘serializeProperties’ from ASIPMessageSerializerHelper to ASIPMessageSerializer
